+++ clients/extension/src/inpage/constants.ts
@@ -1,0 +1,4 @@
+export enum NetworkKey {
+  MAINNET = 'mainnet',
+  TESTNET = 'testnet',
+}
